requirements.txt
playwright
bs4
pandas
chromadb
openai
tiktoken
python-dotenv